moved DatetimeProperties to Series and made it return a Series to be consistent with panads
